Added an agent list as requested
