+++ scripts/Business.js
@@ -7,34 +7,3 @@ export const Business = (businessObject) => {
             <hr class="solid">
     `
 };
-
-export const newYork = (newYorkObject) => {
-    return `
-    <article class="businessList--newYork">
-        <h2 class="business__name">${newYorkObject.companyName}</h2>
-        <div class="business__fullStreet">${newYorkObject.addressFullStreet}</div>
-        <div class="business__cityStateZip">${newYorkObject.addressCity}, ${newYorkObject.addressStateCode} ${newYorkObject.addressZipCode}</div>
-        <hr class="solid">
-    </article>
-    `
-}
-
-export const manufacturing = (manufacturingObject) => {
-    return `
-    <article class="businessList--manufacturing">
-        <h2 class="business__name">${manufacturingObject.companyName}</h2>
-        <div class="business__fullStreet">${manufacturingObject.addressFullStreet}</div>
-        <div class="business__cityStateZip">${manufacturingObject.addressCity}, ${manufacturingObject.addressStateCode} ${manufacturingObject.addressZipCode}</div>
-        <hr class="solid">
-    </article>
-    `
-}
-
-export const purchasingAgent = (purchasingAgentObject) => {
-    return `
-    <article class="agents">
-        <h2 class="business__name">${purchasingAgentObject}</h2>
-        <hr class="solid">
-    </article>
-    `
-}
